refactor(xcode): extract refresh workflow
